added post route for saving activity, renamed private.routes to api.routes
