+++ routes/api.routes.js
@@ -1,13 +1,8 @@
 const router = require("express").Router();
-// const UserModel = require("../models/User.model");
-// const EventModel = require("../models/Event.model");
 const MoodModel = require("../models/Mood.model");
-// const ActivityModel = require("../models/Activity.model");
+// const EventModel = require("../models/Event.model");
+const ActivityModel = require("../models/Activity.model");
 
-//router expects a get from frontend in combination with axios.get
-
-//router expects an post from frontend in combination with axios.post to update database
-//   /api not needed, because defined in index.routes.js
 router.post("/mood", async (req, res) => {
   try {
     console.log(req.body);
@@ -30,6 +25,20 @@ router.get("/moods", async (req, res) => {
     const userId = req.payload._id;
     const moods = await MoodModel.find({ user: userId });
     res.status(200).json({ moods });
+  } catch (err) {
+    console.error("ERROR while getting selected user data :>>", err);
+    res.status(500).json({ message: "Internal Server Error" });
+  }
+});
+
+router.post("/activity", async (req, res) => {
+  try {
+    console.log(req.body);
+    const { title } = req.body;
+    const savedActivity = await ActivityModel.create({
+      title,
+    });
+    res.status(200).json({ savedActivity });
   } catch (err) {
     console.error("ERROR while getting selected user data :>>", err);
     res.status(500).json({ message: "Internal Server Error" });
